♻️ Move `SourceFileResolver` to `extension` package and update imports
Relocated `SourceFileResolver` from `parser` to `extension` package for
improved module organization and updated affected imports and tests
accordingly.
