+++ techdebt-gradle-plugin/src/main/kotlin/com/escodro/techdebt/gradle/extension/SourceFileResolver.kt
@@ -1,4 +1,4 @@
-package com.escodro.techdebt.gradle.parser
+package com.escodro.techdebt.gradle.extension
 
 import java.io.File
